fix: fixed the addEvent Eslint suspense error add: notification can be archived now
