+++ src/app/dashboard/(notif)/NotificationList.tsx
@@ -1,12 +1,31 @@
 import { Notification } from "@/types/notification";
 import { dateToTimeString } from "@/utils/dateUtil";
 import CrossIcon from "@/svgs/CrossIcon";
+import { archiveNotification } from "@/server-actions/notification";
+import { useAlert } from "@/app/(alert)/AlertProvider";
 
 export default function NotificationList({
   notifications,
 }: {
   notifications: Notification[];
 }) {
+  const { showAlert } = useAlert();
+  const handleArchive = async (id: string) => {
+    const res = await archiveNotification(id);
+    if (res.success) {
+      showAlert({
+        title: "Notification archived",
+        description: "",
+        type: "success",
+      });
+    } else {
+      showAlert({
+        title: res.error || "Failed to archive notification",
+        description: "",
+        type: "error",
+      });
+    }
+  };
   return (
     <div className="flex flex-col py-3 border-zinc-600 w-full justify-between items-center overflow-y-auto h-full min-h-0">
       {notifications.map((n: Notification) => (
@@ -18,7 +37,10 @@ export default function NotificationList({
           <div className="flex flex-col w-40 border-l px-3 py-2 border-zinc-600 min-w-40">
             <p>{dateToTimeString(new Date(n.createdAt))} </p>
             <p>{new Date(n.createdAt).toDateString()}</p>
-            <button className="bg-zinc-800 hover:bg-zinc-700 active:bg-red-800 p-2 px-4 rounded-lg flex justify-center items-center">
+            <button
+              className="bg-zinc-800 hover:bg-zinc-700 active:bg-red-800 p-2 px-4 rounded-lg flex justify-center items-center"
+              onClick={() => handleArchive(n.id)}
+            >
               <CrossIcon size={15} />
             </button>
           </div>
