+++ src/app/dashboard/(notif)/UpcomingEventList.tsx
@@ -15,10 +15,10 @@ export default function UpcomingEventList({
           key={e.id}
           className="flex w-full border-b border-zinc-600 items-center justify-between py-2"
         >
-          <p className="p-1 rounded px-7 py-2">
+          <div className="flex flex-col p-1 rounded px-7 py-2 gap-2">
             <Link href={`/dashboard/events/${e.id}`}>{e.title}</Link>
             <UpcomingTag from={e.from} />
-          </p>
+          </div>
           <div className="flex flex-col w-40 border-l px-3 py-2 border-zinc-600">
             <p className="">From: {dateToTimeString(e.from)}</p>
             {e.to && <p className="">To: {dateToTimeString(e.to)}</p>}
